Remove trace events from dispatcher. Move trace to common
