+++ controller/handler_ctrl/trace.go
@@ -5,7 +5,7 @@ import (
 	"github.com/openziti/channel/v2"
 	"github.com/openziti/channel/v2/trace/pb"
 	"github.com/openziti/fabric/common/pb/ctrl_pb"
-	"github.com/openziti/fabric/trace"
+	"github.com/openziti/fabric/common/trace"
 	"google.golang.org/protobuf/proto"
 )
